DG-488 - Merge "core" folder content into "boostrap.go" and move it next to "main.go" (#11)
