+++ main_bootstrap.go
@@ -1,4 +1,4 @@
-/*	
+/*
  *    This file is part of Disgo library.
  *
  *    The Disgo library is free software: you can redistribute it and/or modify
@@ -13,8 +13,8 @@
  *
  *    You should have received a copy of the GNU General Public License
  *    along with the Disgo library.  If not, see <http://www.gnu.org/licenses/>.
-*/
-package core
+ */
+package main
 
 import (
 	"bytes"
@@ -30,11 +30,59 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
+	"sync"
 
-	"github.com/dispatchlabs/disgo/commons/utils"
-	log "github.com/sirupsen/logrus"
+	"github.com/dispatchlabs/disgo/commons/services"
 	"github.com/dispatchlabs/disgo/commons/types"
+	"github.com/dispatchlabs/disgo/commons/utils"
+	"github.com/dispatchlabs/disgo/dapos"
+	"github.com/dispatchlabs/disgo/disgover"
+	"github.com/dispatchlabs/disgo/dvm"
+	log "github.com/sirupsen/logrus"
 )
+
+const (
+	Version = "1.0.0"
+)
+
+// Server -
+type Server struct {
+	services []types.IService
+}
+
+// NewServer -
+func NewServer() *Server {
+	utils.InitializeLogger()
+
+	// Load Keys
+	if _, _, err := loadKeys(); err != nil {
+		utils.Error("unable to keys: " + err.Error())
+	}
+
+	return &Server{}
+}
+
+// Go
+func (server *Server) Go() {
+	utils.Info("booting Disgo v" + Version + "...")
+
+	// Add services.
+	server.services = append(server.services, dvm.GetDVMService())
+	server.services = append(server.services, services.GetDbService())
+	server.services = append(server.services, disgover.GetDisGoverService().WithGrpc().WithHttp())
+	server.services = append(server.services, dapos.GetDAPoSService().WithGrpc().WithHttp())
+	server.services = append(server.services, services.GetHttpService())
+	server.services = append(server.services, services.GetGrpcService())
+
+	// Run services.
+	var waitGroup sync.WaitGroup
+	for _, service := range server.services {
+		utils.Info("starting " + utils.GetStructName(service) + "...")
+		go service.Go(&waitGroup)
+		waitGroup.Add(1)
+	}
+	waitGroup.Wait()
+}
 
 // Keys Helpers
 // ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~ ~~~~
